+++ src/main/java/com/demo/ViNaStack/Model/Symptom.java
@@ -2,25 +2,25 @@ package com.demo.ViNaStack.Model;
 
 import jakarta.persistence.*;
 
-import java.util.ArrayList;
-import java.util.List;
+// import java.util.ArrayList;
+// import java.util.List;
 
 @Entity
 public class Symptom {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String name;
 
     public Symptom() {
 
     }
 
-    public long getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -32,7 +32,7 @@ public class Symptom {
         this.name = name;
     }
 
-    public Symptom(String name, int id) {
+    public Symptom(String name, Long id) {
         this.name = name;
         this.id = id;
     }
